+++ EcoMarket/src/main/java/EcoMarket/EcoMarket/Usuarios/service/EmpleadoService.java
@@ -22,14 +22,36 @@ public class EmpleadoService {
     }
 
     public Empleado ObtenerPorId(int id) {
-        return empleadoRepository.findById(id).orElse(null);
+        return empleadoRepository.findById(id).orElseThrow(() -> new IllegalArgumentException("Empleado no encontrado con ID: " + id));
     }
 
     public Empleado guardar(Empleado empleado) {
+
+        if (empleado.getUsuario() == null) {
+            throw new IllegalArgumentException("El empleado debe tener un usuario asociado.");
+        }
+
+        if (empleado.getNombre() == null || empleado.getNombre().trim().isEmpty()) {
+            throw new IllegalArgumentException("El nombre no puede estar vacío.");
+        }
+
+        if (empleado.getCorreo() == null || empleado.getCorreo().trim().isEmpty()) {
+            throw new IllegalArgumentException("El correo no puede estar vacío.");
+        }
+
+        if (empleado.getCargo() == null || empleado.getCargo().trim().isEmpty()) {
+            throw new IllegalArgumentException("El cargo no puede estar vacío.");
+        }
+
         return empleadoRepository.save(empleado);
     }
 
     public void eliminar(int id) {
+
+        if (!empleadoRepository.existsById(id)) {
+            throw new IllegalArgumentException("No se puede eliminar. Atributo con ID " + id + " no existe.");
+        }
+
         empleadoRepository.deleteById(id);
     }
