+++ EcoMarket/src/main/java/EcoMarket/EcoMarket/Usuarios/service/RolService.java
@@ -21,14 +21,28 @@ public class RolService {
     }
 
     public Rol ObtenerPorId(int id) {
-        return rolRepository.findById(id).orElse(null);
+        return rolRepository.findById(id).orElseThrow(() -> new IllegalArgumentException("Rol no encontrado con ID: " + id));
     }
 
     public Rol guardar(Rol rol) {
+
+        if (rol.getPermisos() == null || rol.getPermisos().isEmpty()) {
+            throw new IllegalArgumentException("El rol debe tener un Permiso.");
+        }
+
+        if (rol.getNombre() == null || rol.getNombre().trim().isEmpty()) {
+            throw new IllegalArgumentException("El Nombre no puede estar vacío.");
+        }
+
         return rolRepository.save(rol);
     }
 
     public void eliminar(int id) {
+
+        if (!rolRepository.existsById(id)) {
+            throw new IllegalArgumentException("No se puede eliminar. Rol con ID " + id + " no existe.");
+        }
+
         rolRepository.deleteById(id);
     }
 }
