+++ EcoMarket/src/main/java/EcoMarket/EcoMarket/Usuarios/service/UsuarioService.java
@@ -22,14 +22,40 @@ public class UsuarioService {
     }
 
     public Usuario ObtenerPorId(int id) {
-        return usuarioRepository.findById(id).orElse(null);
+        return usuarioRepository.findById(id).orElseThrow(() -> new IllegalArgumentException("Usuario no encontrado con ID: " + id));
     }
 
     public Usuario guardar(Usuario usuario) {
+
+        if (usuario.getNombre() == null || usuario.getNombre().trim().isEmpty()) {
+            throw new IllegalArgumentException("El Nombre no puede estar vacío.");
+        }
+
+        if (usuario.getApellido() == null || usuario.getApellido().trim().isEmpty()) {
+            throw new IllegalArgumentException("El Apellido no puede estar vacío.");
+        }
+
+        if (usuario.getEmail() == null || usuario.getEmail().trim().isEmpty()) {
+            throw new IllegalArgumentException("El Email no puede estar vacío.");
+        }
+
+        if (usuario.getContrasena() == null || usuario.getContrasena().trim().isEmpty()) {
+            throw new IllegalArgumentException("El Contrasena no puede estar vacío.");
+        }
+
+        if (usuario.getRol() == null) {
+            throw new IllegalArgumentException("El usuario debe tener un Rol.");
+        }
+
         return usuarioRepository.save(usuario);
     }
 
     public void eliminar(int id) {
+
+        if (!usuarioRepository.existsById(id)) {
+            throw new IllegalArgumentException("No se puede eliminar. Atributo con ID " + id + " no existe.");
+        }
+
         usuarioRepository.deleteById(id);
     }
